refactor(web): extract <MfaPanel> shared by /admin/account-security + /settings/profile (#2257)
Both pages owned an identical block of passkey state + handlers + tile
composition after #2255 shipped /settings/profile. The MFA stack now
lives in a single component so future MFA-flow changes ship to both
surfaces at once.

The new component owns:
- passkey list state, listError, scroll ref
- refreshPasskeys callback + onMount load
- TOTP / passkey change handlers wired to session.refetch
- BackupMethodBanner → PasskeyTile → TwoFactorSetup → BackupCodesStatus
  → PasskeyList → TrustedDevicesList stack
- reauthRedirectTo prop (admin → "/admin/account-security", profile →
  "/settings/profile") forwarded to PasskeyTile

Net: ~50 lines of duplicated state-management removed; -166 / +122 lines
across the three files.
