+++ packages/web/src/app/settings/profile/page.tsx
@@ -1,82 +1,26 @@
 "use client";
 
 /**
- * Per-user settings (no admin role required). MFA tiles reuse
- * /admin/account-security components — the underlying endpoints are auth-only,
- * not admin-gated, so a feature added there shows up here for free.
+ * Per-user settings (no admin role required). MFA tiles reuse the shared
+ * <MfaPanel /> with /admin/account-security — the underlying endpoints
+ * are auth-only, not admin-gated, so a feature added there shows up here
+ * for free.
  */
 
-import { useCallback, useEffect, useRef, useState } from "react";
 import { authClient } from "@/lib/auth/client";
-import { getPasskeyClient, type Passkey } from "@/lib/auth/passkey-client";
-import { TwoFactorSetup } from "@/ui/components/admin/security/two-factor-setup";
-import { PasskeyTile } from "@/ui/components/admin/security/passkey-tile";
-import { PasskeyList, type PasskeyRow } from "@/ui/components/admin/security/passkey-list";
-import { BackupCodesStatus } from "@/ui/components/admin/security/backup-codes-status";
-import { TrustedDevicesList } from "@/ui/components/admin/security/trusted-devices-list";
-import { BackupMethodBanner } from "@/ui/components/admin/security/backup-method-banner";
 import { SectionHeading } from "@/ui/components/admin/compact";
+import { MfaPanel } from "@/ui/components/admin/security/mfa-panel";
 import { IdentitySection } from "@/ui/components/settings/identity-section";
 import { PasswordSection } from "@/ui/components/settings/password-section";
 import { SessionsSection } from "@/ui/components/settings/sessions-section";
 
 interface SessionUser {
   email: string;
-  role: string;
-  twoFactorEnabled?: boolean;
 }
 
 export default function ProfilePage() {
   const session = authClient.useSession();
   const user = (session.data?.user ?? null) as SessionUser | null;
-  const totpEnabled = user?.twoFactorEnabled === true;
-
-  const [passkeys, setPasskeys] = useState<PasskeyRow[] | null>(null);
-  const [listError, setListError] = useState<string | null>(null);
-  const passkeyTileRef = useRef<HTMLDivElement | null>(null);
-
-  function handleEnrollSecondPasskey() {
-    passkeyTileRef.current?.scrollIntoView({ behavior: "smooth", block: "center" });
-  }
-
-  const refreshPasskeys = useCallback(async () => {
-    const client = getPasskeyClient();
-    if (!client) {
-      setListError("Passkey support couldn't be loaded. Refresh the page and try again.");
-      return;
-    }
-    setListError(null);
-    let result: Awaited<ReturnType<typeof client.listUserPasskeys>>;
-    try {
-      result = await client.listUserPasskeys();
-    } catch (err) {
-      const msg = err instanceof Error ? err.message : String(err);
-      console.warn("[passkey] listUserPasskeys threw", msg);
-      setListError("Could not load your passkeys. Please refresh.");
-      return;
-    }
-    if (result.error) {
-      console.warn("[passkey] listUserPasskeys failed", result.error);
-      setListError(result.error.message ?? "Could not load your passkeys.");
-      return;
-    }
-    setPasskeys((result.data ?? []) as Passkey[]);
-  }, []);
-
-  useEffect(() => {
-    void refreshPasskeys();
-  }, [refreshPasskeys]);
-
-  const hasPasskey = (passkeys?.length ?? 0) > 0;
-
-  function handleTotpChange() {
-    session.refetch?.();
-  }
-
-  function handlePasskeyChange() {
-    void refreshPasskeys();
-    session.refetch?.();
-  }
 
   if (!user) {
     return (
@@ -109,32 +53,7 @@ export default function ProfilePage() {
             title="Multi-factor authentication"
             description="Add a second factor — passkey, authenticator app, or both — so a stolen password isn't enough."
           />
-          <div className="space-y-4">
-            <BackupMethodBanner onAddPasskey={handleEnrollSecondPasskey} />
-
-            <div ref={passkeyTileRef}>
-              <PasskeyTile
-                hasPasskey={hasPasskey}
-                onChange={handlePasskeyChange}
-                reauthRedirectTo="/settings/profile"
-              />
-            </div>
-
-            <TwoFactorSetup enabled={totpEnabled} onChange={handleTotpChange} />
-
-            <BackupCodesStatus totpEnabled={totpEnabled} hasPasskey={hasPasskey} />
-
-            <div className="pt-2">
-              <PasskeyList passkeys={passkeys ?? []} onChange={handlePasskeyChange} />
-              {listError && (
-                <p className="mt-2 text-sm text-destructive">{listError}</p>
-              )}
-            </div>
-
-            <div className="pt-2">
-              <TrustedDevicesList />
-            </div>
-          </div>
+          <MfaPanel reauthRedirectTo="/settings/profile" />
         </section>
 
         <SessionsSection />
